Move dependencies into sub-package
