+++ buildSrc/src/main/kotlin/io/spine/gradle/internal/Test.kt
@@ -1,0 +1,57 @@
+/*
+ * Copyright 2021, TeamDev. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.gradle.internal
+
+import io.spine.gradle.internal.dependency.Guava
+import io.spine.gradle.internal.dependency.JUnit
+import io.spine.gradle.internal.dependency.Slf4J
+import io.spine.gradle.internal.dependency.Truth
+
+object Test {
+    const val guavaTestlib = Guava.testLib
+    @Suppress("unused")
+    val truth = Truth
+
+    @Suppress("unused")
+    val junit = JUnit
+    const val junit4 = JUnit.legacy
+
+    @Suppress("unused")
+    @Deprecated("Please do not use.")
+    const val mockito = "org.mockito:mockito-core:2.12.0"
+
+    @Suppress("unused")
+    @Deprecated("Please use Google Truth instead.")
+    const val hamcrest = "org.hamcrest:hamcrest-all:1.3"
+
+    @Deprecated(
+        "Use Flogger over SLF4J.",
+        replaceWith = ReplaceWith("Flogger.Runtime.systemBackend")
+    )
+    @Suppress("DEPRECATION", "unused")
+    const val slf4j = Slf4J.jdk14
+}
